Removed interface class, added GLM and SDL
Having a function set would be better than having a overarching class. 
Moved the RV enum to the operations namespace

The RV enum will probably be used by many other libraries, so it makes no sense for the enum to be in the graphics specific namespace
Modified standard headers

Added the GLM library as the default math library.
Added SDL2 cmake configuration files